Update README.md
The dataset has been provided in an excel sheet in CSV format. The project was done on Google Collaboratory using basics of Machine Learning with Python.
A PDF has been attached which helps in understanding the project better as it includes the problem statement, tools which were used, why Linear Regression was used,
algorithm explained and a conclusion statement.